pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telegram-mcp-complete"
version = "1.0.0"
description = "Complete Telegram MCP Server with QR Code Login"
authors = [{name = "User"}]
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.8.0",
    "telethon>=1.39.0",
    "python-dotenv>=1.1.0",
    "flask>=3.0.0",
    "playwright>=1.40.0",
    "aiohttp>=3.9.0",
    "nest-asyncio>=1.6.0",
]

[project.scripts]
telegram-mcp = "main:main"

[tool.setuptools]
py-modules = ["main", "web_login", "session_manager"]

[tool.black]
line-length = 99
target-version = ['py311']
